react-langchain/requirements.txt
langchain==0.0.352
openai==1.6.1
black==23.12.1
python-dotenv==0.21.0